remove warning

+++ test/vint_test.cpp
@@ -1527,7 +1527,7 @@ CYBOZU_TEST_AUTO(SECP256k1)
 	cybozu::XorShift rg;
 	for (size_t i = 0; i < 100; i++) {
 		Unit x[N * 2];
-		for (int j = 0; j < CYBOZU_NUM_OF_ARRAY(x); j++) {
+		for (size_t j = 0; j < CYBOZU_NUM_OF_ARRAY(x); j++) {
 			x[j] = rg();
 		}
 		compareMod(x, p);
